+++ src/batou/secrets/tests/test_secrets.py
@@ -12,7 +12,13 @@ import tempfile
 
 FIXTURE = os.path.join(os.path.dirname(__file__), "fixture")
 cleartext_file = os.path.join(FIXTURE, "cleartext.cfg")
-encrypted_file = os.path.join(FIXTURE, "encrypted.cfg")
+FIXTURE_ENCRYPTED_CONFIG = os.path.join(FIXTURE, "encrypted.cfg")
+
+
+@pytest.fixture(scope="function")
+def encrypted_file(tmpdir):
+    """Provide a temporary copy of the encrypted confi."""
+    return shutil.copy(FIXTURE_ENCRYPTED_CONFIG, tmpdir)
 
 
 @pytest.fixture(scope="session", autouse=True)
@@ -30,7 +36,7 @@ def cleanup_gpg_sockets():
             pass
 
 
-def test_error_message_no_gpg_found():
+def test_error_message_no_gpg_found(encrypted_file):
     c = EncryptedFile(encrypted_file)
     c.GPG_BINARY_CANDIDATES = ["foobarasdf-54875982"]
     with pytest.raises(RuntimeError) as e:
@@ -40,23 +46,21 @@ def test_error_message_no_gpg_found():
         "`foobarasdf-54875982`")
 
 
-def test_decrypt():
+def test_decrypt(encrypted_file):
     with EncryptedConfigFile(encrypted_file) as secrets:
-        secrets.read()
         with open(cleartext_file) as cleartext:
             assert (cleartext.read().strip() ==
                     secrets.main_file.cleartext.strip())
 
 
-def test_caches_cleartext():
+def test_caches_cleartext(encrypted_file):
     with EncryptedConfigFile(encrypted_file) as secrets:
-        secrets.read()
         secrets.main_file.cleartext = "[foo]bar=2"
         secrets.read()
         assert secrets.main_file.cleartext == "[foo]bar=2"
 
 
-def test_decrypt_missing_key(monkeypatch):
+def test_decrypt_missing_key(monkeypatch, encrypted_file):
     monkeypatch.setitem(os.environ, "GNUPGHOME", '/tmp')
 
     with pytest.raises(batou.GPGCallError):
@@ -65,7 +69,7 @@ def test_decrypt_missing_key(monkeypatch):
         f.read()
 
 
-def test_write_should_fail_unless_write_locked():
+def test_write_should_fail_unless_write_locked(encrypted_file):
     with EncryptedConfigFile(encrypted_file) as secrets:
         secrets.main_file.cleartext = """\
 [batou]
@@ -89,77 +93,71 @@ def test_open_nonexistent_file_for_write_should_create_template_file():
     tf.close()  # deletes file
     encrypted = EncryptedConfigFile(tf.name, write_lock=True)
     with encrypted as secrets:
-        assert not secrets.main_file.cleartext
-        secrets.read()
         assert secrets.main_file.cleartext == NEW_FILE_TEMPLATE
         # The file exists, because we set the write lock
         assert os.path.exists(secrets.main_file.encrypted_filename)
+    # When saving a config without members, delete the file.
+    assert not os.path.exists(secrets.main_file.encrypted_filename)
 
 
-def test_write_unparsable_raises_error():
-    with tempfile.NamedTemporaryFile(prefix="new_encrypted.") as tf:
-        shutil.copy(encrypted_file, tf.name)
-        encrypted = EncryptedConfigFile(tf.name, write_lock=True)
-        with encrypted as secrets:
-            secrets.main_file.cleartext = "some new file contents\n"
-            with pytest.raises(configparser.Error):
-                secrets.read()
+def test_write_unparsable_raises_error(encrypted_file):
+    encrypted = EncryptedConfigFile(encrypted_file, write_lock=True)
+    with encrypted as secrets:
+        secrets.main_file.cleartext = "some new file contents\n"
+        with pytest.raises(configparser.Error):
+            secrets.read()
 
 
-def test_write():
-    with tempfile.NamedTemporaryFile(prefix="new_encrypted.") as tf:
-        shutil.copy(encrypted_file, tf.name)
-        encrypted = EncryptedConfigFile(tf.name, write_lock=True)
-        with encrypted as secrets:
-            secrets.main_file.cleartext = """\
+def test_write(encrypted_file):
+    encrypted = EncryptedConfigFile(encrypted_file, write_lock=True)
+    with encrypted as secrets:
+        secrets.main_file.cleartext = """\
 [batou]
 members = batou
 [asdf]
 x = 1
 """
-            secrets.read()
-            secrets.write()
+        secrets.read()
+        secrets.write()
 
-        with open(encrypted_file, "rb") as old:
-            with open(tf.name, "rb") as new:
-                assert old.read() != new.read()
-        assert 0 != os.stat(tf.name).st_size
+    with open(FIXTURE_ENCRYPTED_CONFIG, "rb") as old:
+        with open(encrypted_file, "rb") as new:
+            assert old.read() != new.read()
+    assert 0 != os.stat(encrypted_file).st_size
 
 
-def test_write_fails_without_recipients():
-    with tempfile.NamedTemporaryFile(prefix="new_encrypted.") as tf:
-        shutil.copy(encrypted_file, tf.name)
-        encrypted = EncryptedConfigFile(tf.name, write_lock=True)
-        with encrypted as secrets:
-            secrets.main_file.cleartext = """\
+def test_write_fails_without_recipients(encrypted_file):
+    encrypted = EncryptedConfigFile(encrypted_file, write_lock=True)
+    with encrypted as secrets:
+        secrets.main_file.cleartext = """\
 [batou]
 members =
 [asdf]
 x = 1
 """
-            secrets.read()
-            with pytest.raises(ValueError):
-                secrets.write()
+        secrets.read()
+        with pytest.raises(ValueError) as exc:
+            secrets.write()
+
+        assert str(exc.value).startswith('Need at least one recipient.')
 
 
-def test_write_fails_if_recipient_key_is_missing_keeps_old_file():
-    with tempfile.NamedTemporaryFile(prefix="new_encrypted.") as tf:
-        shutil.copy(encrypted_file, tf.name)
-        encrypted = EncryptedConfigFile(tf.name, write_lock=True)
-        with encrypted as secrets:
-            secrets.read()
-            secrets.main_file.cleartext = """\
+def test_write_fails_if_recipient_key_is_missing_keeps_old_file(
+        encrypted_file):
+    encrypted = EncryptedConfigFile(encrypted_file, write_lock=True)
+    with encrypted as secrets:
+        secrets.main_file.cleartext = """\
 [batou]
 members = foobar@example.com
 [asdf]
 x = 1
 """
-            secrets.read()
-            with pytest.raises(RuntimeError):
-                secrets.write()
-        with open(tf.name, "rb") as tf_h:
-            with open(encrypted_file, "rb") as encrypted_h:
-                assert tf_h.read() == encrypted_h.read()
+        secrets.read()
+        with pytest.raises(RuntimeError):
+            secrets.write()
+    with open(encrypted_file, "rb") as new:
+        with open(FIXTURE_ENCRYPTED_CONFIG, "rb") as old:
+            assert new.read() == old.read()
 
 
 def test_secrets_override_without_interpolation(tmpdir):